# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: zboard API
  version: v0.0.1
  description: |
    zboard single-binary operations console API for subscriber and operator workflows.
    Business APIs use /api/v1. Zero's standard event webhook is exposed at /api/zero/events.
servers:
  - url: http://localhost:8080
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/users:
    get:
      tags:
        - admin
      summary: List users
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: is_admin
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
            maxLength: 128
        - { name: paged, in: query, schema: { type: boolean }, description: Return the canonical page envelope instead of the legacy array. }
        - { name: sort, in: query, schema: { type: string, enum: [id, email, created_at], default: created_at } }
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        oneOf:
                          - type: array
                            items: { $ref: "#/components/schemas/User" }
                          - $ref: "#/components/schemas/UserPage"
    post:
      tags:
        - admin
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreateRequest"
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/User"
...
  /api/v1/admin/orders:
    get:
      tags: [orders, admin]
      summary: List orders across all users
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 128 }, description: Match order/provider trade number, plan, SKU, channel or a numeric order/user/subscription ID. }
        - { name: status, in: query, schema: { type: string, enum: [attention, pending, paid, failed, canceled, success] }, description: "Use attention for the administrator work queue containing pending and failed orders." }
        - { name: order_type, in: query, schema: { type: string, enum: [new, renewal, upgrade, traffic_pack] } }
        - { name: user_id, in: query, schema: { type: integer, format: int64 } }
        - { name: created_from, in: query, schema: { type: string, format: date }, description: "Inclusive order creation start date. Must be paired with created_to; the window cannot exceed 366 days." }
        - { name: created_to, in: query, schema: { type: string, format: date }, description: "Inclusive order creation end date. Must be paired with created_from." }
        - { name: paged, in: query, schema: { type: boolean }, description: Return the canonical page envelope instead of the legacy array. }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Global order list or page when paged=true
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        oneOf:
                          - type: array
                            items: { $ref: "#/components/schemas/Order" }
                          - $ref: "#/components/schemas/OrderPage"
        "403": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/subscriptions:
    get:
      tags: [subscriptions, admin]
      summary: List subscriptions across all users
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 128 }, description: Match user email, plan, SKU or a numeric subscription/user/plan ID. }
        - { name: status, in: query, schema: { type: string, enum: [active, expired, canceled] } }
        - { name: quota, in: query, schema: { type: string, enum: [available, exhausted] } }
        - { name: user_id, in: query, schema: { type: integer, format: int64 } }
        - { name: expires_from, in: query, schema: { type: string, format: date }, description: "Inclusive subscription expiration start date. Must be paired with expires_to; the window cannot exceed 366 days." }
        - { name: expires_to, in: query, schema: { type: string, format: date }, description: "Inclusive subscription expiration end date. Must be paired with expires_from." }
        - { name: paged, in: query, schema: { type: boolean }, description: Return the canonical page envelope instead of the legacy array. }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Global subscription list or page when paged=true
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        oneOf:
                          - type: array
                            items: { $ref: "#/components/schemas/Subscription" }
                          - $ref: "#/components/schemas/SubscriptionPage"
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    NodeConnectorBearer:
      type: http
      scheme: bearer
      bearerFormat: ZeroNodeAPIKey
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true
        error:
          $ref: "#/components/schemas/ApiErrorDetail"
        timestamp:
          type: string
    ApiErrorDetail:
      type: object
      description: Versioned machine-readable detail included on HTTP error responses. The top-level message remains for backward compatibility.
      required: [version, code]
      properties:
        version:
          type: integer
          enum: [1]
        code:
          type: string
          example: validation_failed
        fields:
          type: object
          description: Field-path to localized actionable message. Present when code is validation_failed.
          additionalProperties:
            type: string
    PageMetadata:
      type: object
      required: [offset, limit, total, next_cursor, previous_cursor]
      properties:
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, format: int64, minimum: 0 }
        next_cursor: { type: string, nullable: true }
        previous_cursor: { type: string, nullable: true }
    PageEnvelope:
      type: object
      required: [page, aggregates, facets]
      properties:
        page: { $ref: "#/components/schemas/PageMetadata" }
        aggregates: { type: object, additionalProperties: true }
        facets: { type: object, additionalProperties: true }
        total: { type: integer, format: int64, minimum: 0, deprecated: true, description: Compatibility alias for page.total. }
        offset: { type: integer, minimum: 0, deprecated: true, description: Compatibility alias for page.offset. }
        limit: { type: integer, minimum: 1, maximum: 200, deprecated: true, description: Compatibility alias for page.limit. }
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        is_admin:
          type: boolean
        status:
          type: string
...
    AdminUserListItem:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [active_subscription_count, total_subscription_count, pending_order_count, total_order_count, created_at]
          properties:
            active_subscription_count: { type: integer, format: int64, minimum: 0 }
            total_subscription_count: { type: integer, format: int64, minimum: 0 }
            pending_order_count: { type: integer, format: int64, minimum: 0 }
            total_order_count: { type: integer, format: int64, minimum: 0 }
            created_at: { type: string, format: date-time }
    UserPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/AdminUserListItem" }
...
    Order:
      type: object
      description: Bounded list projection. Raw payment callback bodies, provider references and failure diagnostics are omitted.
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        subscription_id:
          type: integer
          format: int64
          nullable: true
        plan_id:
          type: integer
          format: int64
        plan_sku_id:
          type: integer
          format: int64
        trade_no:
          type: string
        order_type:
          type: string
          enum: [new, renewal, upgrade, traffic_pack]
        target_subscription_id:
          type: integer
          format: int64
          nullable: true
        amount_cents:
          type: integer
          format: int64
        status:
          type: string
        currency:
          type: string
        channel:
          type: string
        plan_name: { type: string }
        sku_name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    OrderPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Order" }
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        plan_id:
          type: integer
          format: int64
        plan_sku_id:
          type: integer
          format: int64
        start_at:
          type: string
        end_at:
          type: string
        status:
          type: string
        flow_total:
          type: integer
          format: int64
        flow_used:
          type: integer
          format: int64
        node_group_id: { type: integer, format: int64 }
        subscription_type: { type: integer }
        speed_limit_mbps: { type: integer }
        device_limit: { type: integer }
        family_limit: { type: integer }
        renewal_price_minor: { type: integer, format: int64 }
        reset_policy: { type: integer, minimum: 0, maximum: 5 }
        next_reset_at: { type: string, format: date-time, nullable: true }
        traffic_calc_mode: { type: integer, enum: [0, 1, 2] }
    SubscriptionPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/AdminSubscriptionSummary" }
    AdminSubscriptionSummary:
      allOf:
        - $ref: "#/components/schemas/Subscription"
        - type: object
          description: Bounded administration-list projection; subscription config is intentionally omitted.
          properties:
            user_email: { type: string, format: email }
            plan_name: { type: string }
            sku_name: { type: string }
...
    AdminUserCreateRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
        is_admin:
          type: boolean
        status:
          type: string
...